requirements.txt
streamlit==1.30.0
pandas==2.2.1
plotly==5.18.0
google-cloud-bigquery==3.19.0
python-dotenv==1.0.0
